Run post-install-cmd and not full composer install
Running composer install takes time, and if we do it here during run we will need
to deal with time issues outside of containers as docker-compose up will finsih
before this is doen (async), hence uneeded complexity.

So we rather change to instruct how people can do this manually if needed, and if
they have already installed project they won't have to wait for this.